requirements-ocr.txt
# OCR Microservice Dependencies
paddleocr>=2.7.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
